requirements.txt
certifi==2018.11.29
chardet==3.0.4
cycler==0.10.0
GroupyAPI==0.10.2
idna==2.8
kiwisolver==1.0.1
matplotlib==3.0.2
numpy==1.15.4
pandas==0.23.4
pprint==0.1
pyparsing==2.3.0
python-dateutil==2.7.5
python-dotenv==0.10.1
pytz==2018.7
requests==2.21.0
scipy==1.2.0
seaborn==0.9.0
six==1.12.0
urllib3==1.24.1
